refactor: split cli logic from guardoni
