+++ .release-it.js
@@ -38,8 +38,6 @@ module.exports = {
     release: true,
     assets: [
       'extension/dist/*.zip',
-      // ignore extension build for guardoni
-      '!extension/dist/guardoni-*.zip',
       'guardoni/dist/*.deb',
       'guardoni/dist/*.exe',
       'guardoni/dist/*.dmg',
